+++ Triangle.h
@@ -9,7 +9,7 @@ namespace  shapes
 	{
 	public:
 		Triangle() : Polygon() {}
-		Triangle(int x1, int y1, int x2, int y2, double penWidth, COLORREF color, COLORREF fillColor)
+		Triangle(int x1, int y1, int x2, int y2, int penWidth, COLORREF color, COLORREF fillColor)
 			: Polygon(x1, y1, x2, y2, penWidth, color, fillColor)
 		{
 			setPointsNumber(3);
